fix: remove linter warnings
