+++ src/app/components/breadcrumbs/constants.ts
@@ -1,0 +1,9 @@
+export const breadcrumbNameMap: Record<string, string> = {
+  courses: 'RS School',
+  nodejs: 'Node.js Course',
+  'javascript-mentoring-program': 'JavaScript Mentoring Program',
+  'javascript-preschool': 'JavaScript Pre-school',
+  angular: 'Angular Course',
+  'aws-cloud-developer': 'AWS Cloud Developer Course',
+  'aws-fundamentals': 'AWS Fundamentals Course',
+};
