+++ src/app/components/title/types.ts
@@ -1,0 +1,6 @@
+export enum TitleType {
+  Regular = 'regular',
+  Big = 'big',
+  Bigger = 'bigger',
+  ExtraBig = 'extra-big',
+}
