+++ src/app/hooks/use-data-by-name/use-data-by-name.ts
@@ -23,7 +23,7 @@ export const useDataByName = <K extends keyof DataMap>(
       }
     };
     fetchData();
-  }, [dataName]);
+  }, [dataName, fetchDataFn]);
 
   return { data, loading, error };
 };
