+++ src/features/home/alumni/constants.ts
@@ -1,0 +1,42 @@
+import { AlumniProps } from './alumni';
+import aesoft from '@/assets/alumni/aesoft.webp';
+import andersen from '@/assets/alumni/andersen.webp';
+import coherent from '@/assets/alumni/coherent.webp';
+import dataart from '@/assets/alumni/dataart.webp';
+import dott from '@/assets/alumni/dott.webp';
+import elinext from '@/assets/alumni/elinext.webp';
+import epam from '@/assets/alumni/epam.webp';
+import godel from '@/assets/alumni/godel.webp';
+
+import itechart from '@/assets/alumni/itechart.webp';
+
+import miro from '@/assets/alumni/miro.webp';
+import nanosoft from '@/assets/alumni/nanosoft.webp';
+import oxagile from '@/assets/alumni/oxagile.webp';
+import pandadoc from '@/assets/alumni/pandadoc.webp';
+import qulix from '@/assets/alumni/qulix.webp';
+import satellite from '@/assets/alumni/satellite.webp';
+import sberbank from '@/assets/alumni/sberbank.webp';
+import toptal from '@/assets/alumni/toptal.webp';
+import visualfabriq from '@/assets/alumni/visualfabriq.webp';
+
+export const alumni: AlumniProps[] = [
+  { id: 'epam', image: epam },
+  { id: 'toptal', image: toptal },
+  { id: 'oxagile', image: oxagile },
+  { id: 'dott', image: dott },
+  { id: 'andersen', image: andersen },
+  { id: 'godel', image: godel },
+  { id: 'satellite', image: satellite },
+  { id: 'itechart', image: itechart },
+  { id: 'pandadoc', image: pandadoc },
+  { id: 'dataart', image: dataart },
+  { id: 'coherent', image: coherent },
+  { id: 'elinext', image: elinext },
+  { id: 'miro', image: miro },
+  { id: 'qulix', image: qulix },
+  { id: 'visualfabriq', image: visualfabriq },
+  { id: 'sberbank', image: sberbank },
+  { id: 'nanosoft', image: nanosoft },
+  { id: 'aesoft', image: aesoft },
+];
