+++ src/features/home/school/constants.ts
@@ -1,0 +1,19 @@
+import { OptionItemProps } from '@/app/components/option-item/option-item';
+
+export const studyOptions: OptionItemProps[] = [
+  {
+    title: 'Teach and empower',
+    description:
+      'We aim to provide education to everyone interested and assist in hiring the best students.',
+  },
+  {
+    title: 'Connect and grow',
+    description:
+      'We strive to connect people, grow together, and build a solid educational community.',
+  },
+  {
+    title: 'Foster mentoring culture',
+    description:
+      'We encourage a culture of mentoring where knowledge is shared and wisdom is passed on.',
+  },
+];
