+++ src/features/home/school/school.tsx
@@ -1,33 +1,9 @@
-import {
-  OptionItem,
-  OptionItemProps,
-  Paragraph,
-  SectionLabel,
-  Subtitle,
-  Title,
-} from '@/app/components';
+import { studyOptions } from './constants';
+import { OptionItem, Paragraph, SectionLabel, Subtitle, Title } from '@/app/components';
 
 import image from '@/assets/rs-school.webp';
 
 import './school.scss';
-
-export const studyOptions: OptionItemProps[] = [
-  {
-    title: 'Teach and empower',
-    description:
-      'We aim to provide education to everyone interested and assist in hiring the best students.',
-  },
-  {
-    title: 'Connect and grow',
-    description:
-      'We strive to connect people, grow together, and build a solid educational community.',
-  },
-  {
-    title: 'Foster mentoring culture',
-    description:
-      'We encourage a culture of mentoring where knowledge is shared and wisdom is passed on.',
-  },
-];
 
 export const School = () => (
   <div id="school" className="school container">
