+++ src/features/principles/principles.tsx
@@ -1,29 +1,7 @@
-import { PrincipleCard, type PrincipleCardProps, Title, TitleType } from '@/app/components';
-
-import { OpenSourcePhilosophyIcon, OpenToEveryoneIcon, TeachItForwardIcon } from '@/icons';
+import { cards } from './constants';
+import { PrincipleCard, Title, TitleType } from '@/app/components';
 
 import './principles.scss';
-
-export const cards: PrincipleCardProps[] = [
-  {
-    title: 'Open to everyone',
-    description:
-      'Free courses, no obligations, and no contracts. No age limit. Only students’ time and dedication are required. Students can repeatedly attend courses.',
-    icon: <OpenToEveryoneIcon />,
-  },
-  {
-    title: 'Open source philosophy',
-    description:
-      'Our Learning Management System platform and educational materials are publicly available on GitHub and YouTube.',
-    icon: <OpenSourcePhilosophyIcon />,
-  },
-  {
-    title: '"Teach it forward"',
-    description:
-      'Students study at school for free, but we request that they return as mentors to pass on their knowledge to the next generation of students.',
-    icon: <TeachItForwardIcon />,
-  },
-];
 
 export const Principles = () => (
   <div className="principles container">
